test: add uniswap Accuracy test
